% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{.rtractor_palettes}
\alias{.rtractor_palettes}
\title{Rtractor colour palette}
\format{
A named list of character vectors, each containing hex colour codes.
}
\usage{
.rtractor_palettes
}
\description{
Named list of all palettes available in the Rtractor package.
}
\keyword{internal}
