% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rtractor_palettes}
\alias{rtractor_palettes}
\title{List all available Rtractor palettes}
\usage{
rtractor_palettes()
}
\value{
A named integer vector of palette lengths, invisibly.
}
\description{
Prints the names and sizes of all palettes defined in the package.
}
\examples{
rtractor_palettes()

}
