% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_rtractor_c}
\alias{scale_fill_rtractor_c}
\title{Rtractor continuous fill scale}
\usage{
scale_fill_rtractor_c(palette = "seq_blue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name. Default \code{"seq_blue"}.}

\item{reverse}{Logical. Reverse the palette? Default \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_colour_gradientn]{ggplot2::scale_colour_gradientn()}}.}
}
\description{
Interpolates across an Rtractor palette for continuous \code{fill} data.
}
